+++ src/ast/parser.rs
@@ -7,7 +7,7 @@ use crate::{
         BinaryOperator, Expression, FunctionDefinition, Identifier, Program, Statement,
         UnaryOperator,
     },
-    lexer::Token,
+    lexer::{self, Token},
 };
 
 // ques: should i do a trait for 'from(tokens: Vec<Token>) -> ParseResult<Self>'?
@@ -127,35 +127,7 @@ impl Expression {
         let mut next_token = tokens.clone().next().unwrap();
 
         // ques: isn't this too expensive?
-        let is_binary_op = |t: &Token| {
-            matches!(
-                t, // ques: remove clone?
-                Token::Plus
-                    | Token::Negate
-                    | Token::Multiply
-                    | Token::Divide
-                    | Token::Remainder
-
-                    // bitwise operators
-                    | Token::BitwiseAnd
-                    | Token::BitwiseOr
-                    | Token::BitwiseXor
-                    | Token::LeftShift
-                    | Token::RightShift
-
-                    // logical operators
-                    | Token::And
-                    | Token::Or
-
-                    // relational operators
-                    | Token::Equal
-                    | Token::NotEqual
-                    | Token::GreaterThan
-                    | Token::LessThan
-                    | Token::GreaterThanOrEqual
-                    | Token::LessThanOrEqual
-            )
-        };
+        let is_binary_op = |t: &Token| lexer::binary_operators().contains(t);
         while is_binary_op(next_token) && precedence(next_token) >= min_prec {
             debug!(
                 "Found <binop>: {:?} with precedence {}",
@@ -222,19 +194,15 @@ fn precedence(token: &Token) -> i32 {
         Token::Multiply | Token::Divide | Token::Remainder => 50,
         Token::Plus | Token::Negate => 45,
         Token::LeftShift | Token::RightShift => 44,
-        // Token::BitwiseAnd => 43,
-        // Token::BitwiseXor => 42,
-        // Token::BitwiseOr => 40,
         Token::LessThan
         | Token::LessThanOrEqual
         | Token::GreaterThan
         | Token::GreaterThanOrEqual => 35,
-        // Token::BitwiseXor => 31,
         Token::Equal => 30,
         Token::NotEqual => 30,
-        Token::BitwiseAnd => 24,
-        Token::BitwiseXor => 23,
-        Token::BitwiseOr => 22,
+        Token::BitwiseAnd => 22,
+        Token::BitwiseXor => 21,
+        Token::BitwiseOr => 20,
         Token::And => 10,
         Token::Or => 5,
         _ => 0,
