+++ src/codegen/x64/mod.rs
@@ -1,3 +1,4 @@
 pub mod asm;
-pub mod pipe;
 pub mod emit;
+pub mod from;
+pub mod pipe;
